cr feedback - rename EntityInsertPanel - pull 'loading...' out as a constant
